Update: Add menu buttons to all pages
