+++ components/Menu/index.module.css
@@ -1,0 +1,45 @@
+.dropbtn {
+  background-color: var(--mediumseagreen);
+  color: white;
+  padding: 16px;
+  font-size: 1rem;
+  border: none;
+  cursor: pointer;
+}
+
+.dropbtn:hover {
+  background-color: var(--darkseagreen);
+}
+
+.dropdown {
+  position: relative;
+  display: inline-block;
+  margin-right: 30px;
+}
+
+.dropdownContent {
+  position: absolute;
+  background-color: #f1f1f1;
+  min-width: 100px;
+  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
+  z-index: 1;
+  border-radius: 5px;
+}
+
+.dropdownItem {
+  color: var(--darkgray);
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+  border: none;
+}
+
+.dropdownItem:hover {
+  color: var(--gray);
+}
+
+@media screen and (max-width: 640px) {
+  .dropdown {
+    margin-right: 15px;
+  }
+}
